fix 200 test, and moved bootstrap to the route for lightclient.
