+++ data/beaconrestapi/src/test/java/tech/pegasys/teku/beaconrestapi/handlers/v1/beacon/lightclient/GetLightClientBootstrapTest.java
@@ -11,30 +11,33 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.beaconrestapi.handlers.v1.beacon;
+package tech.pegasys.teku.beaconrestapi.handlers.v1.beacon.lightclient;
 
+import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
 import static tech.pegasys.teku.infrastructure.http.HttpStatusCodes.SC_BAD_REQUEST;
 import static tech.pegasys.teku.infrastructure.http.HttpStatusCodes.SC_INTERNAL_SERVER_ERROR;
 import static tech.pegasys.teku.infrastructure.http.HttpStatusCodes.SC_NOT_ACCEPTABLE;
 import static tech.pegasys.teku.infrastructure.http.HttpStatusCodes.SC_NOT_FOUND;
 import static tech.pegasys.teku.infrastructure.http.HttpStatusCodes.SC_NOT_IMPLEMENTED;
+import static tech.pegasys.teku.infrastructure.http.HttpStatusCodes.SC_OK;
+import static tech.pegasys.teku.infrastructure.restapi.MetadataTestUtil.getResponseStringFromMetadata;
 import static tech.pegasys.teku.infrastructure.restapi.MetadataTestUtil.verifyMetadataErrorResponse;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
+import com.google.common.io.Resources;
 import java.io.IOException;
+import java.nio.charset.StandardCharsets;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import tech.pegasys.teku.beaconrestapi.AbstractMigratedBeaconHandlerTest;
+import tech.pegasys.teku.infrastructure.unsigned.UInt64;
+import tech.pegasys.teku.spec.SpecMilestone;
+import tech.pegasys.teku.spec.TestSpecFactory;
+import tech.pegasys.teku.spec.datastructures.lightclient.LightClientBootstrap;
+import tech.pegasys.teku.spec.datastructures.metadata.ObjectAndMetaData;
+import tech.pegasys.teku.spec.util.DataStructureUtil;
 
 public class GetLightClientBootstrapTest extends AbstractMigratedBeaconHandlerTest {
-
-  /*
-  @SuppressWarnings("HidingField")
-  private final Spec spec = TestSpecFactory.createMinimalAltair();
-
-  @SuppressWarnings("HidingField")
-  private final DataStructureUtil dataStructureUtil = new DataStructureUtil(spec);
-   */
 
   @BeforeEach
   void setup() {
@@ -45,20 +48,21 @@ public class GetLightClientBootstrapTest extends AbstractMigratedBeaconHandlerTe
 
   @Test
   void metadata_shouldHandle200() throws IOException {
-    /*
-     LightClientBootstrap lightClientBootstrap =
-         dataStructureUtil.randomLightClientBoostrap(UInt64.ONE);
-     ObjectAndMetaData<LightClientBootstrap> responseData =
-             new ObjectAndMetaData<>(lightClientBootstrap, SpecMilestone.ALTAIR, false, true);
+    final DataStructureUtil dataStructureUtil =
+        new DataStructureUtil(TestSpecFactory.createMinimalAltair());
 
-     final String data = getResponseStringFromMetadata(handler, SC_OK, responseData);
-     final String expected =
-         Resources.toString(
-             Resources.getResource(
-                 GetLightClientBootstrapTest.class, "getLightClientBootstrap.json"),
-             StandardCharsets.UTF_8);
-     assertThat(data).isEqualTo(expected);
-    */
+    LightClientBootstrap lightClientBootstrap =
+        dataStructureUtil.randomLightClientBoostrap(UInt64.ONE);
+    ObjectAndMetaData<LightClientBootstrap> responseData =
+        new ObjectAndMetaData<>(lightClientBootstrap, SpecMilestone.ALTAIR, false, true);
+
+    final String data = getResponseStringFromMetadata(handler, SC_OK, responseData);
+    final String expected =
+        Resources.toString(
+            Resources.getResource(
+                GetLightClientBootstrapTest.class, "getLightClientBootstrap.json"),
+            StandardCharsets.UTF_8);
+    assertThat(data).isEqualTo(expected);
   }
 
   @Test
